DCS-1908: ♻️ Refactor `additionalConditions` into their own sub folder
