+++ server/services/config/conditions/additionalConditions/v1/conditions.ts
@@ -1,6 +1,6 @@
 import moment from 'moment'
-import { ConditionMetadata } from '../../../../data/licenceClientTypes'
-import { AdditionalConditions, AdditionalConditionsV1 } from '../../../../data/licenceTypes'
+import { ConditionMetadata } from '../../../../../data/licenceClientTypes'
+import { AdditionalConditions, AdditionalConditionsV1 } from '../../../../../data/licenceTypes'
 
 enum GroupName {
   PEOPLE_CONTACT_AND_RELATIONSHIPS = 'People, contact and relationships',
